:recycle: self refactoring
